fullstack-login-system/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0
python-dotenv>=1.0.0
# AI / RAG dependencies
openai>=1.30.0
faiss-cpu>=1.8.0
numpy>=1.26.0
tiktoken>=0.7.0
python-multipart>=0.0.9
